Anti-affinity method to set vNIC backing devs
Introduce pypowervm.tasks.sriov and its first method,
set_vnic_back_devs.  This method accepts a VNIC wrapper and candidate
lists of VIOSes and SRIOV physical ports.  It attempts to create backing
devices on the VNIC wrapper such that they are distributed across the
VIOSes and physical SRIOV adapters as much as possible.  See the
docstring for (much) more detail.

Change-Id: I816747d02a9dcf70e78083cb11b3bd0902a5cad4