SR-IOV Logical Ports geared toward SEA
The initial use case for individual SR-IOV logical ports (as opposed to
vNICs) is to be used through Shared Ethernet Adapter.  In that scenario,
the logical port must be set to promiscuous mode, and allowed VLANs/MACs
can't be tailored.

Accordingly, this change set removes the allowed_vlans @property
getter/setter from SRIOVEthLPort.

Change-Id: I13e8c01c6a275ef0f627f8ab1eb7e2db4b9812d8