Add VeNCrypt (TLS/x509) Security to the VNC Repeater
Currently the VNC Repeater is used to be able to open up the LinuxVNC
sessions (started by mkvterm) on a port for remote access.  However,
there is no authentication or encryption done at all on these
connections, so anyone that has access through the firewall can
connect to these sessions and also possibly have access to the
unencrypted data sent across.

This changeset will provide support for optionally specifying
an x509 certificate/key that will be used for x509/TLS security
authentication between the client and the VNC Repeater to allow
for authentication and encrypted transmission.  The VNC Repeater
will continue to communicate with the LinuxVNC session locally
without authentication or encryption.

If a x509 certificate/key is specified, the VNC Repeater will
enforce that the client negotiates to use VeNCrypt as part of
the security handshake and then will wrapper the socket to use
TLS 1.2 for secure communication.

Change-Id: I8c4eb1dbe365697ab3592c6528e1c7df41814701