Make convert_bytes_to_gb honor dp properly
The doc for pypowervm.util.convert_bytes_to_gb states that, if dp is
specified, the result is rounded *up* to that number of decimal places.
However, it also states that it will return low_value if the result
would be smaller than low_value.

In the case where dp is less than the number of decimal places in
low_value, the behavior was inconsistent: if the input value initially
converted to a number smaller than low_value, the method would return
low_value - ignoring dp.  For example:

convert_bytes_to_gb(1, dp=2)  ==>  0.0001

This change set corrects the behavior (and clarifies the doc) such that:

convert_bytes_to_gb(1, dp=2)  ==>  0.01

Change-Id: I56ae977e8875af5867f049ae8ec373a2542577b1
Closes-Bug: 1599591