Implement WrapperElemList.__repr__
Although there is no hope of complying with the (loosely-suggested)
contract that eval(repr(object)) == object, the fact that different
tools (e.g.  ipython, pdb) use different means to display
representations of objects can hamper usability.

In this instance, it was observed that pdb uses repr when evaluating a
symbol.  Because WrapperElemList didn't provide an implementation for
__repr__, the result was always '[]' (the default from the list
superclass).

This change set implements __repr__ on WrapperElemList to display
something useful, roughly equivalent to what __str__ does.

Change-Id: Idfe9bd456aea4ee4d31145993b38d68fb52fb6dd