Move code from logical_partition to base_partition
Move checks for whether or a not a capability is modifiable to
the parent class. This allows all partitions to call these methods.

The methods that were migrated are located in a new class
and added to BasePartition as a mixin.

Managment partitions will no longer be declared as migratable.
Change-Id: I4bf125876dfdc3f7e644a177ff5118ad071c3d9f