Move code from logical_partition to base_partition
Move checks for whether or a not a capability is modifiable to
the parent class. This allows VIOS, or managment partitions to
call these methods.

The new methods that were moved, will be located in a new file
and used as a mixin to reduce complexity of the BasePartition class.
2 classes are also moved to the new file, then imported to
base_partition.py to avoid changing consumers of this file.

Change-Id: I4bf125876dfdc3f7e644a177ff5118ad071c3d9f