Handle port insertion for stale vfc mapping
Because of stale and/or rebuilt VFC mappings, the Port
element of a VirtualFibreChannelMapping may be missing. When adding a
map for a virtual wwpn pair that already matches an existing mapping,
we insert the Port element if it is missing.

Also, when deriving the npiv map,
we add support for the case where we don't want to preserve
an existing mapping that matches. This can happen if we are rebuilding
the maps for stale physical port scenario which can happen if we detach
an NPIV volume, migrate the VM, and then attempt to attach a volume
again. Instead of just checking for a non existent backing_port like
the add_map method, it was thought better to control with an optional
'preserve' boolean parameter.

Change-Id: Ia8553f6fe4e9e590bc5de7259e076c84ddd966b1
Partial-Bug: #1585624