Optimize calls for get_x_partition methods
This change puts a small optimization in for the partition methods.  It
will only call searches on the LPARs if they're absolutely required.
This serves as a good optimization for these queries if the mgmt
partition is of type VIOS.  If the mgmt partition is type LPAR, then the
algorithm is no worse than before.

VIOSes are checked first because it is highly likely that there are
fewer of those in the system than VIOSes.

It removes the check for multiple responses as that is not something
that the REST API should ever return.

Change-Id: Iec3ac4f07766e64efcccc249611b392aa212e1f5