Close vterm when user navigates away
Before this change, when a user closes a VNC session, the backing vterm
and VNC stick around.  This can be a high memory overhead, because each
can take multiple megabytes of memory.

The simple fix for this would be to detect when a user closes the client
side VNC we should close the Server VNC and vterm.  However, that fully
closes out the vterm session.  If a user navigated away for a few
moments and then re-opened a whole new session would be created and the
original context could be lost.  This was not ideal.

This change is similar to the above.  It will close the vterm when the
user navigates away from the VNC server.  However, the key difference
is that it will keep the original VNC session open for 5 minutes.

If they navigate away from the VNC for more than 5 minutes it will close
out the VNC Server session and the backing vterm.

This strikes a good balance between making sure too many VNC's don't
take up all the memory on the NovaLink while maintaining good usability
within the session.

This also changes the select from 10 second time out to 1 second to
allow for quicker killing of sessions.

Change-Id: I2626ae80fa505ba8c4d1ad32b72051ac39c337bb