Close vterm when user navigates away
This change will close the vterm when the user navigates away from the
VNC server.  It will keep the original VNC session open for 5 minutes.
If they navigate away from the VNC for more than 5 minutes it will close
out the VNC session.  This allows for the use case where the user closes
the VNC (maybe they go to a different link in a noVNC session) and then
come back later.

If it waits too long, it closes the session.  Otherwise it preserves it.
This strikes a good balance between making sure too many VNC's don't
take up all the memory on the NovaLink while maintaining good usability
within the session.

This also changes the select from 10 second time out to 1 second to
allow for quicker killing of sessions.

Change-Id: I2626ae80fa505ba8c4d1ad32b72051ac39c337bb