Close vterm when user navigates away
This change will close the vterm when the user navigates away from the
VNC server.

QUESTION: Should we say 'close after 5 minutes if no new sessions come
in?'  use case for this, I navigate to a new tab in the browser to get
some data (maybe an IP address of another VM) and then come back.  I
don't want to lose the whole context?

Change-Id: I2626ae80fa505ba8c4d1ad32b72051ac39c337bb