Interpret SRIOV pport granularity as percent * 100
In the REST schema, SR-IOV physical port minimum ethernet capacity
granularity is represented as a long integer which needs to account for
a percentage value to two decimal places.  Hence the percentage value is
multiplied by 100 - e.g. 2% is represented as 200.

We want to return this value as a float between 0 and 1 to match the
other capacity-related getters in the wrappers.  Therefore, we need to
divide the provided long by 10000.

Change-Id: Ic0bbabd0dc13a7e42860e0486eeebe05dc1aa657