SRIOV*PPort backpointer to SRIOVAdapter (and ID)
In working with SR-IOV adapters and physical ports, it was discovered
that developers frequently needed to be able to discover/reference the
port's owning SR-IOV Adapter - usually its sriov_adap_id property.
Previously, code would hack an instance attribute into the SRIOV*PPort
wrapper and subsequent code would rely on that attribute.

This change set adds two new pseudo-@propertys to SRIOVEthPPort (and, by
inheritance, SRIOVConvPPort):
- sriov_adap is a reference to the SRIOVAdapter to which the physical
  port belongs.
- sriov_adap_id is the SRIOVAdapter's sriov_adap_id.

This allows simplification of consuming code that needs to be able to
refer back to the SRIOVAdapter (or its ID) from the context of a
physical port wrapper.

Change-Id: I601a40ab5c0f8ea01efc3e85e60b53024ddda617