Add VNIC slot info to slot_map
With the SR-IOV story comes the new VNIC wrapper. The client slot
numbers and MACs for these adapters need to be stored in the slot map
so that the client adapter MACs on the target side of an evacuation
or rebuild receive the same slot numbers they orginally had on the
source.

On the slot_map store side the schema fits right in with the existing
structure for vSCSI, vFC, and CNA slot data. It will look like the
following for VNICs:

    { slot_num: { 'VNIC': { mac_addr: None } } }

At the time of this change it is determined that no extra data is
needed to be stored with VNICs, so the value keyed by the MAC address
will be None. This _may_ change in the future.

The methods added as part of this commit are generalized to include both
CNAs and VNICs. The term 'vnet' is used to mean either a CNA or VNIC.

Change-Id: I826ccc44a8c529163ecd15553f38f98e5981c22b