add_lpar_storage_scrub_tasks(remove_storage=True)
pypowervm.tasks.storage.add_lpar_storage_scrub_tasks adds FeedTask logic
to delete storage mappings and backing storage elements.  The new
remove_storage kwarg, which defaults to True for backward compatibility,
can be used to restrict the logic to mappings only.  This can be useful
in community flows where storage removal in the destroy flow is
contingent upon the context of the destroy.

Change-Id: I7c00f2f672df64ba7875f3ead223c8db4522a38a