Make pvid optional to VNIC[Details].bld()
It is valid for a vNIC to be created with no PVID, implying its traffic
will be untagged.  This is implemented in REST by omitting the
PortVLANID attribute from the PUT payload, whereupon it is defaulted to
zero by the platform.

This change set renders the pvid argument to VNIC.bld and
VNICDetails.bld optional.  Further, it translates a PVID value of zero
to None to reinforce the semantic that there is no PVID in this case.

Change-Id: Iff23a0f7a215193d4f82409e7ea816b154b2a737