Ease up default log helper level
The default log helper stored 5 of the last API calls.  On scale up
systems this is far too agressive.  It was understood that this may need
to be eased back eventually when it was created.

We now have data showing that easing back the log helper a bit can
reduce the memory load on the system.  This is because the responses are
stored in memory.  This slight tweak could remove 50-100 MB off
processes that are making multiple VIOS calls on a scale up system.  It
also maintains value as the last 3 calls are usually the most important,
rather than the last 5.

Change-Id: I6a04430edc3e68e5eee5302cfe9bbc2f35006e8e
Closes-Bug: bug/1608474