Don't autovivify [Converged]EthernetPhysicalPorts
When an SR-IOV adapter has no EthernetPhysicalPorts (only Converged
ports - or, in the case of an adapter in Dedicated mode, no physical
ports at all), the XML does not contain the <EthernetPhysicalPorts>
property.

Previously, when we went to grab the list of physical ports via the
phys_ports @property getter, the _find_or_seed helper was autovivifying
the <EthernetPhysicalPorts/> property in the XML. This would (probably)
work fine if we were creating it in the right place - but since we don't
have an element order set up for the ManagedSystem or its nested
properties, it can get injected in a random (and schema-incorrect) spot.

This change set switches _find_or_seed to just _find to avoid
autovivifying this property if it doesn't exist.

The above applies to both EthernetPhysicalPorts and
ConvergedEthernetPhysicalPorts.

Change-Id: I79368d672396ca9fa552cc7c2672ef04249ee451
Closes-Bug: 1617050