Add vlan mappings to live migration call
This change set adds the VLAN mappings to the live migration invocation.
This allows for the migration to update the VLAN information on an
ethernet adapter when it finishes its migration.

The VLAN update will occur immediately when the VM resumes on the target
system.  This will be before the migration actually completes.  The
resume on the target means that the system has started running the VM on
the target, but there is still data being transferred from the source
system.

Change-Id: Ifb10858e1bd8ad93065ddc981f108006e8c6f88d