Better Handling of mkvterm error scenarios
Support was previously added to try to handle the error case where
the vterm was started out-of-band without the vnc option, including
the ability to force the previous one closed.  However, the PVME
error message ids relied upon for this are at least no longer being
returned from mkvterm, so all error scenarios were resulting in a
cryptic error saying it couldn't parse an integer from an empty
string.

This change will take a different approach for handling such errors.
First, there is an agreed upon return code of 3 in the situation
where a vterm is already active.  However, there are 2 different
scenarios for this return code of 3.  There is the case where it was
started with the VNC option in which case a port number is returned
and we can use it, and the case where no port number is returned
meaning that it was started without VNC, which we either need to
force closed (if they asked to force), or fail with the actual
reason.  All other return codes of non-0 are scenarios where we
need to fail giving them the actual reason for the failure.

Also, since we may be failing for a reason other than because
of the return code of 3 we don't necessarily want to put the
force explanation in the exception throw.  Which even in the
return code of 3 case is not always helpful, where if the caller
wishes to return the exception as-is, they won't want their
caller to see an explanation of using the force option, when
they have made the choice not to specify force on their call.
So, making a change to the exception being thrown to just
provide the generic terminal session failure with the exception
string.

Change-Id: If22795f25c3d19ee738760add4eeed6215e443d4