Colocate multiple fabrics on VIOSes
The current logic for migration can end with migration mappings like the
following:

Source                  Target
6/vios1/fcs0 (A)    6/vios1/fcs0 (A)
3/vios2/fcs0 (A)    3/vios2/fcs4 (A)
4/vios1/fcs2 (B)    4/vios2/fcs0 (B)
5/vios2/fcs2 (B)    5/vios1/fcs4 (B)

The problem with this is that the last mappings are not correct.  Client
slots 4 and 6 should be paired and 3 and 5 should be paired, even though
they're on different fabrics.

This change set introduces a multi-fabric-aware algorithm.  It should
make sure that the source slots land on the same target slots.

Change-Id: I0f2f002ec02d43aec8ddb5552d5f0c7b47beb06b
(cherry picked from commit 2930c1dcf78d32a066a982bd1519a2d77dfc2ffe)