toxmlstring(pretty)
Add a 'pretty' kwarg to Wrapper.toxmlstring() and Element.toxmlstring().
If True, we pass pretty_print=True to etree.tostring(), causing the XML
to be formatted nicely with newlines and tabs for human viewing, a boon
to logging/debugging.  If False, there is no effect.  (In particular, we
do not pass pretty_print=False to etree.tostring() - being really
careful about backward compatibility.)

Change-Id: I9d1fc41ad0cb0e3e3b68568321796acd5ac69445