Retry the upload on the API in case of failure
Occasionally the API upload (typically associated with vOpt) will fail
to upload with an error about 0 bytes written.  It is unclear where this
issue is happening - either in the REST Server or via the python sockets
library.  Given the general rarity of the failure (one in about a
thousand) and the fact that other uploads can be going on at same time
without issue, retry logic seems prudent.

This change set catches an upload failure and will retry up to three
times.  If the error persists it will raise up the issue to the end
user.

Change-Id: Ia6519d36b706cd70b9dd31e2f9e0abae4160d364
Closes-Bug: 1621730