Implement remove_iscsi
This will make an API call to logout of a specified iSCSI volume.
Logging out of the iSCSI volume will remove it from the compute node.
The method should be called whenever an iSCSI volume is disconnected
from an instance, but ONLY after all related scsi mappings have
been removed from the VIOS.

Change-Id: I9765a9e37840a5b2e738bd621113a64850db7687