WIP: Support LU storage types in RebuildSlotMap
Currently when rebuilding an LPAR and preserving slot assignments for I/O
devices, we only support vSCSI rebuilds that have a PV storage type. In this
change set, the RebuildSlotMap will be expanded to support rebuilding an LPAR
that has a LU storage type. The _vscsi_build_out method will now allow PV and
LU storage types to be initialized in the rebuild map.

Change-Id: I2c66d79d5ea6efa4b7b4999c7a12f7a290b7e0be