WIP: Additional query params on Adapter.read
This change set adds a new add_qp kwarg to pypowervm.adapter.Adapter's
.read, .build_path, and .extend_path methods.  If unspecified, this
kwarg has no effect.  If specified, must be a list of (key, value)
tuples to be added to the query string of the request.  For example:

adap.read('SomeElement', add_qp=[('one', 'two'), ('three', 'four')])

...will result in the request being performed with path
'.../SomeElement?group=None&one=two&three=four'

(Note that the implicit group=None is added because xag was not
specified.  If xag=[] is specified, group=None is absent, but the add_qp
query params are still appended correctly.)

Change-Id: I04fa9deeae8a7027e30895f436ca237fa838129e