set_vnic_backdevs can consider pport link status
Previously, set_vnic_backdevs would merrily use a backing physical port
whose link status was 'down', on the assumption that the user would only
include ports in the candidate port list that she actually wanted to
use.  (If they're down right now, it might be because they're not yet
cabled, but they're going to be soon; or there's a network glitch that's
getting resolved; or whatever.)

With this change set, the user has the option of indicating via a new
kwarg, check_port_status, whether link-down ports should be allowed as
backing devices.  By default, the behavior is the same as before -
link-down ports are used.  If the new kwarg is set to True, link-down
ports are eliminated from consideration as backing devices.

Change-Id: Icd199fa3e2145f57fca429059835392bc5e1108a