Custom Event with VNet device create/update/delete
Today, the platform's Event API has no affordance for creation,
modification, or deletion of the "virtual network device" data types:
ClientNetworkAdapter, SRIOVEthernetLogicalPort, and VirtualNICDedicated.
This means that a consumer trying to detect such an occurrence requires
heavy polls on its parent LogicalPartition.  This problem is compounded
when the consumer doesn't know the parent LPAR.  For example, in a spawn
operation in OpenStack, the neutron agent only knows the MAC address it
expects to see on the VIF as it is plugged (created).

To help reduce the need for this polling, this change set causes the
create, update, and delete methods of these three types to produce a
custom Event which can be consumed by any standard Event listener.  The
Event contains:

    type: CUSTOM_CLIENT_EVENT
    data: <REST URI of the object>
    detail: JSON string representing
            {'action': <'create'|'update'|'delete'>,
             'pvid': <PVID>, 'mac': <MAC address>,
             'orig_pvid' <PVID>,  # update only
             'orig_mac': <MAC address>  # update only
            }

By capturing this Event, a MAC- or PVID-oriented consumer need not poll
for the vnet device object itself (possibly across all LPARs); and in
fact need not even GET the object afterward if the MAC/PVID is the only
thing it needs to know.

Note: Consumers should still listen for Events of types NEW_CLIENT,
CACHE_CLEARED, and MISSING_EVENTS, whereupon the heavy discovery should
still be performed.  (Since this should be rare, the overall load should
still be greatly reduced.)

Change-Id: I500fc52214701ba8bb9e644f7a7c04db3a69b0d4