Get rid of Event polling interval
The pseudo-synchronous behavior of the Event API (described below) makes
it unnecessary and generally undesirable for the client to sleep at all
when polling for events.  This change set removes the interval, causing
the _EventPollThread to spin hard on the Event GET.

The Event API provides pseudo-synchronous behavior as follows:

- If any events are waiting on the server, the GET returns them all,
  immediately.
- If there are no events ready, the server waits for some to arrive.  As
  soon as any events arrive, the GET returns them all.
- If no events happen on the server within 10 seconds, the GET returns
  an empty feed.

This paradigm is specifically designed to allow the client to poll
without sleeping, providing optimal efficiency in receiving events.

Change-Id: Iaf5969ec4a54ed7ccc6d4f76feb58edd0f811d15