Add debug and ensure close on upload
The upload of the stream could potentially not close the output stream
if there was an exception thrown during the input stream read.  This
change adds detection of the error and ensures that close.  It also adds
some debug to the overall code path to make it easier for users to
investigate the progress of the upload.

Change-Id: Id39a4f9ae586bc5df90102cb8fe56424f165c942