Prevent Hang on VNC Handshake Receive
As part of the handshake for the VeNCrypt authentication support
it will do receives and sends back and forth which work fine in
normal scenarios.  However, if the network drops one of the
packets then the recv call may hang forever, and especially
in the single port scenario this is one thread for all VM's
so will hang up everything.

Making a change here such that before each of the recv calls it
will do a select to make sure that data can be received.  It will
use a timeout of 30 seconds and will fail the request if a timeout
occurs (since nothing else to do if a packet is missing).

Change-Id: I8a18fa3785478ecde4840e69472a9e018159e216