Fixes for custom events & wrapper-based events
1) Exception in the Event GET are merrily swallowed by the event
handler; but the resulting lack of initialization of the response would
blow up the recently-added wrapper-based processing, killing the event
handler thread.  With this change set, we only do the wrapping when the
response is really there.

2) In the case of such an exception, since the 15s loop interval was
removed, the event handler would thrash the connection trying to re-GET
the Event feed.  This change set adds a short sleep to the exception
path to wait for the server to recover.

3) Add new event type CUSTOM_CLIENT_EVENT to recognized-but-unhandled
(along with VISIBLE_URI) in _EventListener code that produces legacy raw
and dict-formatted event interpretations.

This is to get rid of error log messages like:

ERROR pypowervm.adapter Unexpected EventType=CUSTOM_CLIENT_EVENT

Change-Id: I7cb52694dfcc1eca3288ec59406c6b028bd7f200