[PATCH] DO NOT MERGE.  Local-to-remote connection
This patch enables the local system to use pypowervm.adapter to talk to
a remote PVM REST server as if it were local.  The patched version of
Session finds a live RMC address and uses that for the pypowervm host IP
address.

Requirements:
- install ConfigParser
- enable 'sudo' to work for the current user without requiring a
  password.  This is needed to get the RMC address.

Configuration:
- If the PVM REST server credentials are nonstandard, create
  pypowervm.ini in the home directory of the user as follows:

[config]
user = neo
password = bmVvMTIz

...where the 'password' value is the base64 encoded version of the
cleartext password (echo -n 'cleartext_password' | base64)

Unit tests and pep8 will not pass with this patch in place.

DO NOT MERGE THIS PATCH.  IT IS INTERNAL ONLY.

Change-Id: I20b5be5e80a521cc98df238aeab836df227664c5