FileIO wrapper
This change set introduces pypowervm.wrappers.storage.FileIO, which is
really just a <VirtualDisk/> whose bld method only accepts a 'label'.
This is to be used by attaching it to a VSCSIMapping which, when PUT to
the REST server, will actually create the File I/O object on the back
end.

We also introduce VDiskType to mirror VirtualDiskType.Enum; and set the
FileIO's VirtualDiskType field to 'File' as appropriate.

Change-Id: Ic0d931259bc0f455e4c2807637e0957a99cb5865