[WIP]: remove_*_mappings can search on UDIDs
Changes in pypowervm.tasks.scsi_mapper to allow gen_match_func and all
of the remove_*_mappings methods to accept UDID-related arguments (list
or single value, as appropriate to the specific method).  The UDID(s)
is/are ignored if any of the name-related params are used.  If the
name-related params are unspecified, we match on the UDID(s).  If none
are specified, we still match all mappings for the specified LPAR.

This behavior is intended to be fully backward compatible.

[WIP]: UT

Change-Id: I2ffd7f08205939dd92b2be64a69616916d849093