Change default wait time for VIOS RMC activity based on time since boot
10 minutes is a long time to wait for RMC activity from VIOSes, and
in the event there's at least one RMC active VIOS we would still
allow platform operations. But this 10 minute max wait may impede
on other caller timeouts.

We want to wait 10 minutes because when a system first boots it may
take this long for VIOS RMC states to go active, and we want to have
active VIOS RMC states for mechanisms that might do something like
mapping cleanup for remote restart after calling validate_vios_ready.

The solution here is to only wait a maximum of 10 minutes if we know
we're running close to boot time. Otherwise we'll wait a maximum of
2 minutes. This should be a fair compromise because if we know a
system has been running for awhile, it's unlikely given 8 extra
minutes that a VIOS's RMC state would go active.

Change-Id: Icc3f2cdb47f448391ac36c5bfbf21003518363b0