_AllowedList.const_or_list
Refactor pypowervm.util._AllowedList (parent class of VLANList and
MACList) to split out the piece that discovers whether one of the const
values was specified.  If so, returns that value; otherwise, returns the
list with its values validated and sanitized as appropriate to the REST
API for the respective subclass type.

Change-Id: Iabf77db1a3dde15ac813b4b4bfea42894a09ef85