Return empty string for mtms.MTMS.mtms_str if substring is NoneType
In the case of Clusters, a Node may come up from REST without a MTMS
attribute. The mtms_str method, in that case, will then raise this error:

"unsupported operand type(s) for +: 'NoneType' and 'str'"

This changeset gets around that issue by checking if any of the attributes
are None, and accordingly returning the empty string if the are.

Change-Id: I24edbc48573c2b396e0702af24a0e06ca73eceb5