Provide function delegate for upload
Previously the only way to upload an element through the PowerVM API
required end users to pass in a stream of data.  However, under the
covers many of those copies just sent the stream to a file path (if
running locally).  Certain consumers found it easier and potentially
faster if they could write the data directly to the path, rather than
wrapping it in the stream.

This change set allows this to occur.  The upload methods now can take
in a method.  That method will be invoked with a single parameter - the
path to where the upload should be sent.  The method will block until an
EOF is sent.

In some cases, this has sped up upload time by 50%.

This change is fully backwards compatible due to the new upload_type
parameter.  It defaults to the original upload types (stream based).  To
use the new options, users will have to explicitly change the upload
type.

Change-Id: Ibd215be5768eab96eddde35eee7b65ac62b893ef