[WIP] Accept */* for quick props
When requested correctly, with valid parameters, quick properties are
returned as JSON.

However, when requested with e.g. a bogus UUID, the payload for the
error (in this case 404) is returned as XML.

Previously, the Adapter code explicitly set the Accept header to
application/json for quick properties.  This meant that any error
condition would get masked by a 406 (Not Acceptable) because the
server wanted to return the real error as application/atom+xml.

This change set makes the Accept header '*/*' for quick properties
so that both scenarios are handled correctly.

(Ideally, we would send both Accept headers - application/json and
application/atom+xml - but the requests library doesn't seem to
provide a way to do that.)

Change-Id: Ie6b52e3ede1672e2597c035abdfb89f578702572
WIP: UT and buy-in