Remove psutil_compat and psutil dependencies
The need for psutil_compat was removed when references to it were
eliminated via
https://github.com/powervm/pypowervm/commit/52532908d82f4a8224d7e0573292010440ffe62e
but psutil_compat itself remained.  This change set removes it and its
test suite.

psutil_compat was the only thing using psutil; so removing it allows us
to remove our dependencies on psutil as well.  This change set removes
the dependencies from requirements.txt and both Debian and RPM packaging
specs.

Change-Id: Ia73c1f8c963341a35e4ee2f1e4ca39163e996ba5