Base VIOS wait time on VIOS uptime
This is a different approach to
b5479740a836181b82edc1f2e2e5577978a29a14.

>>> From b5479740a836181b82edc1f2e2e5577978a29a14
10 minutes is a long time to wait for RMC activity from VIOSes, and
in the event there's at least one RMC active VIOS, we would still
allow platform operations. But this 10 minute max wait may impinge
on other caller timeouts.

We want to wait 10 minutes because when a system first boots it may
take this long for VIOS RMC states to go active, and we want to have
active VIOS RMC states for mechanisms that might do something like
mapping cleanup for remote restart after calling validate_vios_ready.

The solution here is to only wait a maximum of 10 minutes if we know
we're running close to boot time. Otherwise we'll wait a maximum of
2 minutes. This should be a fair compromise because if we know a
system has been running for awhile, it's unlikely given 8 extra
minutes that a VIOS's RMC state would go active.
<<< From b5479740a836181b82edc1f2e2e5577978a29a14

Though it's not clear from the above description, the previous change
was using the NovaLink's uptime as the basis for calculating the default
timeout.  With this change set, we instead use the uptime of each VIOS,
calculate the default timeout on a per-VIOS basis, and time out of the
entire operation only if all VIOSes have passed their respective
timeouts.  This only applies to VIOSes whose power state is active, but
whose RMC state is inactive.

Change-Id: I4a6d909f5f4270f3140d766fef695deb9a6fa51f