PoC Code to handle remote streams piped to local
Pypowervm recently introduced the ability to have clients directly pipe
to a file.  This was cool, but it broke remotability.  This change is a
PoC that abstracts that.  It allows client code to use the local path,
but under the covers uses a named pipe to stream over the remotable REST
API.

Change-Id: Iaf6b8d5c09ee2db549f00ddf218e04a2187c9e44