WIP: Ensure close is sent to FIFO pipe
Pypowervm was recently updated to allow the user to pass in a function
for the streaming to the REST API.  They would be passed the path to the
FIFO pipe.  However, we found situations where the function that was
passed in would raise an exception immediately.  This is acceptable on
the client code (something went wrong), but pypowervm would block
indefinitely.

It blocked indefinitely because an EOF was never written to the REST
API, thus never letting the command finish.  This change fixes that and
wraps the client function in a try/except block.  This ensures that
pypowervm will send an EOF to the FIFO pipe and allow the REST API to
give back control.

TODO: Unit Test
TODO: Live Test

Change-Id: I3ec0119dd58d3f41a5aec3bf62c9b274d0c9a01c
Closes-Bug: 1640953