Delete the vDisk when the upload fails
When a stream is being uploaded into the I/O server, there is the
potential for the stream to fail.  In the event of this, the element
should be deleted and the exception should be raised up.

This change set fixes the vDisk upload issue.  It does not resolve the
issue if it occurs on a LU or vOpt.  That will be handled in a different
change set.

Change-Id: I54fe2a165adf9de16128fe7a340a949e9c998ed3
Partially-Closes: 1641987