Abort the fifo pipe
The FIFO pipe worked on some systems but not others.  This appears to be
due to the GIL from python.  It did not handle the FIFO pipe being both
read and written to from the same process.  Even when switching to
ProcessPoolHandler, the connections failed.

This change provides a quick workaround.  It will allow the
delegate_function to still work.  However it temporarily streams it to
/tmp and then uploads to the REST API.  It is in fact quite similar in
structure to the original fifo code.

In a future release, we may find a way to stream directly, but for now
this is considered a compatability function.  We do not expect that most
users need this function, and instead if working remote will use the
UplaodType of STREAM or STREAM_BUILDER.

Change-Id: I61fc5abe9c0e9a7e1989aa9657b81bcb21fcf1b1