Fix exception stack formatting
The wrapped exceptions in the FeedTask did not provide more than a
single line of detail for what the error was.  It made debug impossible.

The WrappedException that is caught (from TaskFlow) has the original
failure contained within a WrappedException.  The 'fail.exception'
actually just resolved to the exception string, not the full actual
exception.

Running LOG.excption(fail) only prints out the stack trace of the
wrapped exception, not the underlying exception.

This is a quirk of dealing with TaskFlow.  The workaround to get the
full exception trace is fixed here.  This allows for proper root cause
analysis of issues contained within FeedTasks.

Change-Id: Ic10ff5ed4e3842e9b1d3c4e0930f5f489bb02d16
Closes-Bug: 1646556