Add support for remove_optical job parameters to power on job
2 new parameters have been added at the REST layer to the LPAR power on
job to allow for the removal of the virtual optical config drive. This
changeset enables and tests the following job parameters:

remove_optical_time: an integer value specifying how long to wait before
                     deleting the media
remove_optical_name: the name of the VOptMedia to delete

remove_optical_time is optional, and defaults to 0 (i.e. immediate deletion
during the power on job). remove_optical_name is required if a user wishes
to delete a media device.

Change-Id: I54554857b521afaff29e2e392541ae948847885a