Fix storage scrubbers for VIOS-to-VIOS mappings
When the Novalink partition is marked as a VIOS type, its SCSI mappings
were subject to accidental removal because the scrub code only accounted
for LPAR types as VSCSI clients.  This change set fixes the scrubbing
code to account for both VIOS and LPAR VSCSI clients to avoid these
accidental removals.

Change-Id: I613835725c30a2467b4a5b02d661b61501abc056
Closes-Bug: #1627579