Refactor pypowervm.tasks.power
This change set refactors the code for power_off and power_on.  The
legacy behavior through those methods should be (more or less)
preserved.  The module now offers new primitives, PowerOp.start and
PowerOp.stop, which just perform the requested operation, with no
second-guessing or progressive retries.  The main goal of this change
set is to make the code more readable and maintainable.

Change-Id: If6885f6b53724e4361c2ae70048a988c3d0aef38