WIP: Add support for slot_map extra spec query
This change set adds support for FileIO types to be saved in the slot
stor.  However, it also adds support for querying the slot stor by the
extra spec.  It does this because the FileIO type can not pre-determine
the UDID.  However, it does have the path to the file, and can therefore
look up the slot based off the path to the file.

Change-Id: I197dec9903fe665fd5d889dee5784c0ad11d1def