Use a modern PBR package
The 2.0.0 is breaking in that it removes the use of warnerrors in
build_sphinux.

pypowervm isn't using that feature, so it shoudln't break.

The cap on pbr is preventing OpenStack projects that would like to
use pbr 2.0.0 (and sphinx 1.5.1) from doing so as it breaks
co-installability with pypowervm

Related-Bug: https://bugs.launchpad.net/openstack-requirements/+bug/1668848