Rename FileIO.bld to bld_ref
For consistency with other cases where we're not actually PUTting an
object to the REST server, but just creating a reference to an existing
object that can then be used within another element (e.g. VSCSIMapping),
this change renames the artist formerly known as FileIO.bld to
FileIO.bld_ref.  The old name is retained for backward compatibility.

Change-Id: I851ef64d0098c465fad8dd2d3e94f9e55f3a051c