Change to use new RSCT memory metrics if they are available.
Originally, only PctRealMemFree was available, but that counts
cached memory as in use, when in reality it can be freed to be used
by other processes. AIX has additional memory metrics which can
be used to calculate the actual in use memory (TotalPgSpSize,
TotalPgSpFree, VMActivePageCount, and RealMemSize) and Ubuntu has
an additional metric PctRealMemAvailable. These new metrics allow
for calculating VM memory usage without counting cached memory
as in use.

Change-Id: Ia40c385135a28d90841ed6a55bbdd09d17bd195f