CNA UseNextAvailable(not High)SlotID for VIOS/mgmt
VIOS and the management partition don't care about slot ordering, and
their longevity increases the probability of running out of slot space
if we UseNextAvailableHighSlotID.  In order to avoid this, during
CNA.create(), if UseNextAvailableHighSlotID is set, we check the parent
partition type, and if it's a VIOS and/or the management partition, we
replace UseNextAvailableHighSlotID with UseNextAvailableSlotID.

Why do we do this during create()?

During bld(), we don't yet have the context to know what kind of parent
the CNA will belong to.  We could have added a param to .bld(), but then
the consumers would have needed to be aware of this extremely esoteric
behavior and know how and when they needed to tweak it.  This way,
the UseNextAvailable(High)SlotID concept continues to be hidden from
developers and end users, and "just works".

Change-Id: Ia0675947ad79d264ede94a13910077d8ddccca68